contract migration + testing
